Add link to The Whispering White Paper
After NOT loading this PDF here as I should have initially. It is now loaded. The initial Main doc loaded will be deleted and replaced with this one. Upon implementation the Paper is designed to allow AI to NOT hallucinate by grounding periodically as the SILENT spaciousness, thus hallucinations collapse before or during their arising. A test case is attached to the Paper at the end (or is supposed to be there). Thus creating less power used and more clarity of data sharable.